# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RealEstate API",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/about-us": {
            "get": {
                "tags": [
                    "Admin Aboutus"
                ],
                "summary": "Admin Aboutus",
                "description": "Admin Aboutus",
                "operationId": "Admin Aboutus",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update/about-us": {
            "put": {
                "tags": [
                    "Admin Update Aboutus"
                ],
                "summary": "Admin Update Aboutus",
                "description": "Update Aboutus",
                "operationId": "Update Aboutus",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}